nsync works for lulu

+++ src/nsync.bash
@@ -1,4 +1,5 @@
 local="$(hostname)"
+user="$(whoami)"
 binary="unison"
 
 if [[ "$local" == "desktop.palmer.dabbelt.com" ]]
@@ -27,10 +28,35 @@ then
 	mv $temp ~/.kde4/share/config/kopeterc
 fi
 
-rm dead.letter 2> /dev/null
-rm ~/.unison/*.log 2> /dev/null
+if [[ "$local" == "lululi" ]]
+then
+	echo "lululi <==> server.dabbelt.com"
+	
+	ssh server.dabbelt.com nsync
+	ionice -n 7 -c 2 $binary laptop
+	ssh server.dabbelt.com nsync
+fi
 
-pwd=`pwd`
-cd ~/prog/
-make 2> /dev/null > /dev/null
-cd "$pwd"
+if [[ "$user" == "palmer" ]]
+then
+	rm dead.letter 2> /dev/null
+	rm ~/.unison/*.log 2> /dev/null
+
+	pwd=`pwd`
+	cd ~/prog/
+	make 2> /dev/null > /dev/null
+	cd "$pwd"
+fi
+
+if [[ "$user" == "lulu" ]]
+then
+	rm ~/.unison/*.log 2> /dev/null
+	
+	pwd=`pwd`
+	for x in $(/bin/ls ~/.programs/ | grep -v bin)
+	do
+		cd ~/.programs/$x
+		make
+	done
+	cd $pwd
+fi
